reorganize hash header (again) and add hash_range
