--- include/yk/hash.hpp
@@ -1,10 +1,0 @@
-#ifndef YK_HASH_HPP
-#define YK_HASH_HPP
-
-#include "yk/hash/std.hpp"
-
-#if YK_UTIL_INCLUDE_BOOST
-#include "yk/hash/boost.hpp"
-#endif
-
-#endif  // YK_HASH_HPP
